+++ users/marcel/kdreams/jam_io.h
@@ -81,16 +81,16 @@ typedef enum ct_TYPES
 
 struct COMPStruct
 {
-	unsigned long DecompLen;
+	uint32_t DecompLen;
 
 };
 
 
 struct CMP1Header
 {
-	unsigned CompType;					// SEE: ct_TYPES above for list of pos.
-	unsigned long OrginalLen;			// Orginal FileLength of compressed Data.
-	unsigned long CompressLen;			// Length of data after compression (A MUST for LZHUFF!)
+	uint16_t CompType;					// SEE: ct_TYPES above for list of pos.
+	uint32_t OrginalLen;			// Orginal FileLength of compressed Data.
+	uint32_t CompressLen;			// Length of data after compression (A MUST for LZHUFF!)
 };
